reduce failure logging in TSocket::openConnection
TSocket::local_open() calls openConnection for all addresses associated with a given
hostname and succeeds on first one that succeeds. The failed tries lead to unnecessary
socket failure logs, so added a flag to openConnection to skip logging which is
set to true from local_open() only for if the address being tried is the last one.

Removed inExit flag and instead check for existing openSSLInitialized before trying
to close SSL connections.